+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unversioned
+
+- Breaking: Removed majority of, since February 18th 2023, deprecated IRC commands.
+
 ## v5.0.1
 
 - Minor: Added deprecation notices for some client methods.